Split AutoGsonAdapterGenerator into multiple files and improved code
