--- gsonpath-compiler/src/main/java/gsonpath/generator/adapter/AutoGsonAdapterProperties.kt
@@ -1,13 +1,0 @@
-package gsonpath.generator.adapter
-
-import com.google.gson.FieldNamingPolicy
-import gsonpath.GsonFieldValidationType
-import gsonpath.PathSubstitution
-
-internal class AutoGsonAdapterProperties(val fieldsRequireAnnotation: Boolean,
-                                         val flattenDelimiter: Char,
-                                         val serializeNulls: Boolean,
-                                         val rootField: String,
-                                         val gsonFieldValidationType: GsonFieldValidationType,
-                                         val gsonFieldNamingPolicy: FieldNamingPolicy,
-                                         val pathSubstitutions: Array<PathSubstitution>)
