+++ gsonpath-compiler/src/main/java/gsonpath/generator/adapter/standard/AutoGsonAdapterProperties.kt
@@ -1,0 +1,14 @@
+package gsonpath.generator.adapter.standard
+
+import com.google.gson.FieldNamingPolicy
+import gsonpath.GsonFieldValidationType
+import gsonpath.PathSubstitution
+
+internal class AutoGsonAdapterProperties(
+        val fieldsRequireAnnotation: Boolean,
+        val flattenDelimiter: Char,
+        val serializeNulls: Boolean,
+        val rootField: String,
+        val gsonFieldValidationType: GsonFieldValidationType,
+        val gsonFieldNamingPolicy: FieldNamingPolicy,
+        val pathSubstitutions: Array<PathSubstitution>)
