+++ gsonpath-compiler/src/main/java/gsonpath/generator/adapter/standard/AutoGsonAdapterPropertiesFactory.kt
@@ -1,4 +1,4 @@
-package gsonpath.generator.adapter
+package gsonpath.generator.adapter.standard
 
 import com.google.gson.FieldNamingPolicy
 import gsonpath.*
